task-70: add labels to inputs
