+++ src/assets/locales/ru/translation.json
@@ -52,7 +52,7 @@
     },
     "new_event": {
       "title": "Новое мероприятие",
-      "event_name": "Введите название мероприятия",
+      "event_name": "Введите название мероприятия*",
       "event_name_example": "Поход в горы",
       "dates": "Выберите даты",
       "date_start": "Дата начала",
