+++ src/js/view/elements/inputs/InputProps.ts
@@ -3,6 +3,7 @@ import React, { InputHTMLAttributes } from "react";
 export interface InputProps {
   placeholder?: InputHTMLAttributes<HTMLInputElement>["placeholder"];
   value?: InputHTMLAttributes<HTMLInputElement>["value"] | null;
+  label?: string;
   className?: string;
   disabled?: boolean;
   type?: "date" | "email" | "number" | "time";
