+++ src/js/view/elements/inputs/TextArea/TextAreaProps.ts
@@ -3,6 +3,7 @@ import { InputHTMLAttributes } from "react";
 export interface TextAreaProps {
   placeholder?: InputHTMLAttributes<HTMLInputElement>["placeholder"];
   value?: string | null;
+  label?: string;
   className?: string;
   disabled?: boolean;
   rows?: number;
